refactor(widget): split add_widget_page into separate files under add_widget/ folder
