+++ lib/pages/wizard/widget_page.dart
@@ -1,5 +1,5 @@
 import 'package:bugaoshan/l10n/app_localizations.dart';
-import 'package:bugaoshan/pages/settings/add_widget_page.dart';
+import 'package:bugaoshan/pages/settings/add_widget/add_widget_page.dart';
 import 'package:bugaoshan/widgets/common/third_center.dart';
 import 'package:flutter/material.dart';
